More refactoring!
